+++ CHANGELOG.md
@@ -12,7 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### ✨ Added
 
+- HTTPS support for homeassistant LED devices (#1886)
+
 ---
+
 
 ### 🔧 Changed